converter-app/requirements.txt
PyMuPDF==1.24.3
pdf2docx==0.5.8
python-docx==1.1.0
pytesseract==0.3.13
Pillow==10.4.0
layoutparser==0.3.4
customtkinter==5.2.2
tkinterdnd2==0.4.2
PyInstaller==6.9.0
